Add temporary warning to detect schedule-in-between situations
By schedule-in-between I mean situations where we reserve machine for
some time interval T, when it is already reserved for some T' s.t.
T' > T (T'.start >= T.end). If such situation happens it might indicate
that current implementation contains a pretty serious bug when looking
for a machine predecesor of to-be-scheduled operation.