feat(solver,ecdk): allow for CLI & config file control on whether to use local search in fitness evaluation (#190)
## Description <!-- Please describe the motivation & changes introduced
by this PR -->

As title states, it is now possible to specify **on solver's** CLI
whether to use local search operator when computing fitness.
Please note that this option was not yet added to ecdk's CLI, as I'm
hoping to achieve it with #164.

Right now best option to is to pass it with ecdk is via appropriate
config file.

## Linked issues <!-- Please use "Resolves #<issue_no> syntax in case
this PR should be linked to an issue -->

Closes #174