feat(ecdk): add new solution discovery (#296)
## Description <!-- Please describe the motivation & changes introduced
by this PR -->

Check against solution database for newly discovered solutions.

Added creation of sqlite3 database with solution reference, so that I do
not have to load whole reference to RAM.

I'm thinking right now, that this would be possible with polars w/o
usage of sqlite, but I'm not gonna rewrite this.

## Linked issues <!-- Please use "Resolves #<issue_no> syntax in case
this PR should be linked to an issue -->

Closes #251

## Important implementation details <!-- if any, optional section -->